[metagraph.p{xd,yx}] Unsigned int -> size_t

+++ zeroforcing/metagraph.pxd
@@ -4,7 +4,7 @@ from sage.data_structures.bitset_base cimport bitset_t
 from zeroforcing.fastqueue cimport FastQueueForBFS
 
 cdef class ZFSearchMetagraph:
-    cdef unsigned int num_vertices, vertex_to_fill
+    cdef size_t num_vertices, vertex_to_fill
     cdef dict neighbors_dict, closed_neighborhood_list, orig_to_relabeled_verts, relabeled_to_orig_verts
     cdef set vertices_set
     cdef bitset_t meta_vertex
@@ -13,7 +13,7 @@ cdef class ZFSearchMetagraph:
 
     cdef void initialize_neighbors(self, object)
     cdef FrozenBitset extend_closure(self, FrozenBitset, FrozenBitset)
-    cdef void neighbors_with_edges_add_to_queue(self, FrozenBitset, FastQueueForBFS, unsigned int)
+    cdef void neighbors_with_edges_add_to_queue(self, FrozenBitset, FastQueueForBFS, size_t)
 
     @staticmethod
     cdef list shortest(FrozenBitset, FrozenBitset, list, dict)
@@ -21,9 +21,9 @@ cdef class ZFSearchMetagraph:
     cdef set build_zf_set(self, list)
     cpdef set dijkstra(self, frozenset, frozenset)
 
-    cpdef object to_orig_vertex(self, unsigned int)
+    cpdef object to_orig_vertex(self, size_t)
     cpdef object __to_orig_metavertex_iter(self, object)
     cpdef frozenset to_orig_metavertex(self, object)
-    cpdef unsigned int to_relabeled_vertex(self, object)
+    cpdef size_t to_relabeled_vertex(self, object)
     cpdef object __to_relabeled_metavertex_iter(self, object)
     cpdef frozenset to_relabeled_metavertex(self, object)
